Create RedHat deploy configs (#382)

+++ test/deploy/linux/php/templates/install-wordpress.sh
@@ -13,7 +13,19 @@ cd /wordpress
 
 echo "<?php phpinfo(); ?>"| tee test.php
 
-mysql -e \
+# Ensure /usr/local/bin is on the path as it doesn't appear
+# to be, by default, on the Linux 2 image.
+if [[ ":$PATH:" != *:/usr/local/bin:* ]]; then
+	export PATH=$PATH:/usr/local/bin
+fi
+
+# If on RedHat based distro, assume we are using mariadb, and
+# that it needs to be started post package installation.
+if [ $(which yum) ]; then
+	service mariadb start
+fi
+
+mysql -u root -e \
 "GRANT ALL ON *.* TO 'wp_user'@'localhost' IDENTIFIED BY 'wp_password'; \
 FLUSH PRIVILEGES;"
 
@@ -39,17 +51,21 @@ wp core install --allow-root \
             --admin_email=admin@example.com
 
 if [ $(which nginx) ]; then
-	if [ ! -f /etc/nginx/sites-available/wordpress-site ]; then
-		echo Creating nginx wordpress-site config.
-		cp ~/templates/wordpress-site /etc/nginx/sites-available
-	fi
-	if [ ! -L /etc/nginx/sites-enabled/wordpress-site ]; then
-		echo Activating nginx wordpress-site config.
-		ln -s /etc/nginx/sites-available/wordpress-site /etc/nginx/sites-enabled/
-	fi
-	if [ -L /etc/nginx/sites-enabled/default ]; then
-		echo Disabling default nginx site config.
-		unlink /etc/nginx/sites-enabled/default
+	if [ -d /etc/nginx/sites-available ]; then
+		if [ ! -f /etc/nginx/sites-available/wordpress-site ]; then
+			echo Creating nginx wordpress-site config.
+			cp ~/templates/wordpress-site /etc/nginx/sites-available
+		fi
+		if [ ! -L /etc/nginx/sites-enabled/wordpress-site ]; then
+			echo Activating nginx wordpress-site config.
+			ln -s /etc/nginx/sites-available/wordpress-site /etc/nginx/sites-enabled/
+		fi
+		if [ -L /etc/nginx/sites-enabled/default ]; then
+			echo Disabling default nginx site config.
+			unlink /etc/nginx/sites-enabled/default
+		fi
+	else
+		sed -i -e "s/\/usr\/share\/nginx\/html/\/wordpress/" /etc/nginx/nginx.conf
 	fi
 	echo Restarting nginx.
 	systemctl restart nginx
@@ -82,4 +98,12 @@ if [ $(which apache2) ]; then
 
 	echo Restarting apache2.
 	service apache2 reload
+fi
+
+# The presence of httpd indicates we are running on a Linux 2 host.
+# The FPM / no FPM decision is made via packages in this case.
+if [ $(which httpd) ]; then
+	cp ~/templates/wordpress-httpd.conf /etc/httpd/conf.d/
+	sed -i -e "s/DocumentRoot \"\/var\/www\/html/DocumentRoot \"\/wordpress/" /etc/httpd/conf/httpd.conf
+	service httpd restart
 fi
